added function pointers

+++ tests/cases/error_sizeof_function.c
@@ -1,7 +1,4 @@
 /* SPDX-License-Identifier: MIT */
 /* expect-error: invalid application of sizeof to function type */
 int f(void);
-
-int main(void) {
-    return sizeof(f);
-}
+int main(void) { return (int)sizeof(f); }
